migrate changelog to compose resources
