+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/StreetCompleteApplication.kt
@@ -14,7 +14,6 @@ import de.westnordost.streetcomplete.data.CacheTrimmer
 import de.westnordost.streetcomplete.data.CleanerWorker
 import de.westnordost.streetcomplete.data.Preloader
 import de.westnordost.streetcomplete.data.allEditTypesModule
-import de.westnordost.streetcomplete.data.changelog.changelogModule
 import de.westnordost.streetcomplete.data.dbModule
 import de.westnordost.streetcomplete.data.download.downloadModule
 import de.westnordost.streetcomplete.data.download.tiles.DownloadedTilesController
@@ -119,7 +118,6 @@ class StreetCompleteApplication : Application() {
                 mainModule,
                 maptilesModule,
                 metadataModule,
-                changelogModule,
                 noteEditsModule,
                 notesModule,
                 messagesModule,
