--- app/src/androidMain/kotlin/de/westnordost/streetcomplete/data/changelog/ChangelogModule.kt
@@ -1,7 +1,0 @@
-package de.westnordost.streetcomplete.data.changelog
-
-import org.koin.dsl.module
-
-val changelogModule = module {
-    single { Changelog(get()) }
-}
