+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/screens/about/ChangelogScreen.kt
@@ -26,11 +26,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import de.westnordost.streetcomplete.R
+import de.westnordost.streetcomplete.data.changelog.Changelog
 import de.westnordost.streetcomplete.ui.common.BackIcon
 import de.westnordost.streetcomplete.ui.common.HtmlText
 import de.westnordost.streetcomplete.ui.ktx.plus
 import de.westnordost.streetcomplete.ui.theme.titleLarge
-import de.westnordost.streetcomplete.util.html.HtmlNode
 
 /** Shows the full changelog */
 @Composable
@@ -63,7 +63,7 @@ fun ChangelogScreen(
 
 @Composable
 fun ChangelogList(
-    changelog: Map<String, List<HtmlNode>>,
+    changelog: Changelog,
     modifier: Modifier = Modifier,
     paddingValues: PaddingValues = PaddingValues()
 ) {
