+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/screens/about/ChangelogViewModel.kt
@@ -3,24 +3,25 @@ package de.westnordost.streetcomplete.screens.about
 import androidx.compose.runtime.Stable
 import androidx.lifecycle.ViewModel
 import de.westnordost.streetcomplete.data.changelog.Changelog
-import de.westnordost.streetcomplete.util.html.HtmlNode
+import de.westnordost.streetcomplete.data.changelog.readChangelog
 import de.westnordost.streetcomplete.util.ktx.launch
+import de.westnordost.stretcomplete.resources.Res
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 
 @Stable
 abstract class ChangelogViewModel : ViewModel() {
     /* version name -> html */
-    abstract val changelog: StateFlow<Map<String, List<HtmlNode>>?>
+    abstract val changelog: StateFlow<Changelog?>
 }
 
 @Stable
-class ChangelogViewModelImpl(private val changes: Changelog) : ChangelogViewModel() {
-    override val changelog = MutableStateFlow<Map<String, List<HtmlNode>>?>(null)
+class ChangelogViewModelImpl(private val res: Res) : ChangelogViewModel() {
+    override val changelog = MutableStateFlow<Changelog?>(null)
 
     init {
         launch {
-            changelog.value = changes.getChangelog()
+            changelog.value = res.readChangelog()
         }
     }
 }
