+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/messages/Message.kt
@@ -1,8 +1,8 @@
 package de.westnordost.streetcomplete.data.messages
 
+import de.westnordost.streetcomplete.data.changelog.Changelog
 import de.westnordost.streetcomplete.data.user.achievements.Achievement
 import de.westnordost.streetcomplete.data.user.achievements.Link
-import de.westnordost.streetcomplete.util.html.HtmlNode
 
 sealed interface Message {
     /** User has unread messages in his OSM inbox */
@@ -16,7 +16,7 @@ sealed interface Message {
     ) : Message
 
     /** The app was updated to a new version */
-    data class NewVersion(val changelog: Map<String, List<HtmlNode>>) : Message
+    data class NewVersion(val changelog: Changelog) : Message
 
     /** The user is informed about being able to select which quests he wants to solve */
     data object QuestSelectionHint : Message
